% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{create_task_graph}
\alias{create_task_graph}
\title{Create the task graph}
\usage{
create_task_graph(file_info, comp_info, comp_args)
}
\arguments{
\item{file_info}{tibble with file info}

\item{comp_info}{tibble with component info}

\item{comp_args}{tibble with component arguments}
}
\value{
igraph
}
\description{
Create the task graph
}
\examples{
\dontrun{
create_task_graph(file_info, comp_info, comp_args)
}
}
