% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_example.R
\name{.render_example_process_property}
\alias{.render_example_process_property}
\title{process each property with indentation}
\usage{
.render_example_process_property(prop, prop_name = NULL, indent_level = 0)
}
\arguments{
\item{prop}{property}

\item{prop_name}{property name}

\item{indent_level}{indentation level}
}
\value{
array of strings
}
\description{
process each property with indentation
}
\examples{
\dontrun{
.render_example_process_property(prop, prop_name, indent_level)
}
}
