% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{.task_graph_get_root}
\alias{.task_graph_get_root}
\title{get task graph root
#'}
\usage{
.task_graph_get_root(task_api)
}
\arguments{
\item{task_api}{The task api filepath}
}
\value{
string with root of the task graph
}
\description{
get task graph root
#'
}
\examples{
\dontrun{
.task_graph_get_root(task_api)
}
}
