% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_comp_args_as_tibble}
\alias{format_comp_args_as_tibble}
\title{Format the arguments of a component as a tibble}
\usage{
format_comp_args_as_tibble(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
tibble with file info
}
\description{
Format the arguments of a component as a tibble
}
\examples{
\dontrun{
format_comp_args_as_tibble(spec)
}
}
