% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_file_format.R
\name{format_file_format}
\alias{format_file_format}
\title{Format the file spec format}
\usage{
format_file_format(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
string
}
\description{
Format the file spec format
}
\examples{
\dontrun{
format_file_format(spec)
}
}
