% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_file_format_as_kable.R
\name{format_file_format_as_kable}
\alias{format_file_format_as_kable}
\title{Format file spec format as kable}
\usage{
format_file_format_as_kable(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
table
}
\description{
Format file spec format as kable
}
\examples{
\dontrun{
format_file_format_as_kable(spec)
}
}
