% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_as_tibble.R
\name{list_as_tibble}
\alias{list_as_tibble}
\title{set list as tibble}
\usage{
list_as_tibble(li)
}
\arguments{
\item{li}{list}
}
\value{
dataframe of list
}
\description{
set list as tibble
}
\examples{
\dontrun{
list_as_tibble(li)
}
}
