% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tibble.R
\name{list_contains_tibble}
\alias{list_contains_tibble}
\title{check if list contains tibble
#'}
\usage{
list_contains_tibble(li)
}
\arguments{
\item{li}{list}
}
\value{
boolean
}
\description{
check if list contains tibble
#'
}
\examples{
\dontrun{
list_contains_tibble(li)
}
}
