% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openproblems.docs-package.R
\docType{package}
\name{openproblems.docs-package}
\alias{openproblems.docs}
\alias{openproblems.docs-package}
\title{openproblems.docs: OpenProblems Documentation}
\description{
OpenProblems Documentation R helper functions. Deprecated: superseded by the \code{openproblems.project.docs} module of the Python \code{openproblems} package.
}
\section{Deprecated}{

This package is deprecated and will be removed in a future release. The
README rendering pipeline (\code{common/scripts/create_task_readme}) runs the
Python \code{openproblems} package, so \code{openproblems.project.docs} is the source
of truth for task documentation.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://openproblems.bio/documentation}
  \item \url{https://github.com/openproblems-bio/core}
  \item Report bugs at \url{https://github.com/openproblems-bio/core/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood)

Authors:
\itemize{
  \item Kai Waldrant \email{k_waldant@hotmail.com} (\href{https://orcid.org/0009-0003-8555-1361}{ORCID})
}

}
\keyword{internal}
