% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_spec.R
\name{read_anndata_slots}
\alias{read_anndata_slots}
\title{Read anndata slots}
\usage{
read_anndata_slots(spec, path)
}
\arguments{
\item{spec}{file spec}

\item{path}{Path to yaml file}
}
\value{
tibble with file slots
}
\description{
Read anndata slots
}
\examples{
\dontrun{
read_anndata_slots(spec, 'path/to/yaml')
}
}
