% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_api.R
\name{read_comp_spec}
\alias{read_comp_spec}
\title{read component spec
#'}
\usage{
read_comp_spec(path)
}
\arguments{
\item{path}{path to comp yaml}
}
\value{
a list with compontent info and arguments
}
\description{
read component spec
#'
}
\examples{
\dontrun{
read_comp_spec("src/path/to/comp.yaml")
}
}
