% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_spec.R
\name{read_file_info}
\alias{read_file_info}
\title{Read file info}
\usage{
read_file_info(spec, path)
}
\arguments{
\item{spec}{file spec}

\item{path}{Path to yaml file}
}
\value{
tibble with file info
}
\description{
Read file info
}
\examples{
\dontrun{
read_file_info(spec, 'path/to/yaml')
}
}
