% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project_config.R
\name{read_project_config}
\alias{read_project_config}
\title{Read project config}
\usage{
read_project_config(file)
}
\arguments{
\item{file}{Path to a project config file}
}
\description{
Read project config
}
\examples{
file <- system.file("extdata", "example_project", "_viash.yaml", package = "openproblems.docs")

read_project_config(file)
}
