% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_task_api.R
\name{read_task_api}
\alias{read_task_api}
\title{Read the api files in a task}
\usage{
read_task_api(path)
}
\arguments{
\item{path}{Path to a src directory}
}
\value{
A list with the api info
}
\description{
Read the api files in a task
}
\examples{
path <- system.file("extdata", "example_project", package = "openproblems.docs")

read_task_api(path)
}
