% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_task_config.R
\name{read_task_config}
\alias{read_task_config}
\title{Read project config}
\usage{
read_task_config(path)
}
\arguments{
\item{path}{Path to a project config file}
}
\description{
Read project config
}
\section{Deprecated}{

Superseded by \code{openproblems.project.docs.read_task_config()} in the Python
\code{openproblems} package, which is what \code{create_task_readme} runs.
}

\examples{
path <- system.file(
  "extdata", "example_project", "_viash.yaml",
  package = "openproblems.docs"
)

read_task_config(path)
}
