% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_component.R
\name{render_component}
\alias{render_component}
\title{Render component section}
\usage{
render_component(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
string
}
\description{
Render component section
}
\examples{
path <- system.file("extdata", "example_project", "api", "comp_method.yaml", package = "openproblems.docs")

render_component(path)
}
