% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_file.R
\name{render_file}
\alias{render_file}
\title{Render file section}
\usage{
render_file(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
string
}
\description{
Render file section
}
\examples{
\dontrun{
render_file(spec, 'path/to/yaml')
}
}
