% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_file_format.R
\name{render_file_format}
\alias{render_file_format}
\title{Render file section}
\usage{
render_file_format(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
string
}
\description{
Render file section
}
\section{Deprecated}{

Superseded by \code{openproblems.project.docs.render_file_format()} in the Python
\code{openproblems} package, which is what \code{create_task_readme} runs.
}

\examples{
path <- system.file(
  "extdata", "example_project", "api", "file_train_h5ad.yaml",
  package = "openproblems.docs"
)

spec <- read_file_format(path)

render_file_format(spec)
}
