% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_json_schema_example.R
\name{render_json_schema_example}
\alias{render_json_schema_example}
\title{Render an example of a JSON schema as a YAML string}
\usage{
render_json_schema_example(json_schema)
}
\arguments{
\item{json_schema}{JSON schema as a list}
}
\value{
YAML string
}
\description{
This function takes a JSON schema and renders an example YAML string.
}
\section{Deprecated}{

Deprecated along with the rest of this package; it has no replacement in the
Python \code{openproblems} package.
}

\examples{
library(openproblems)
example <- system.file("extdata", "example_schema.json", package = "openproblems.docs")

json_schema <- read_nested_yaml(example)

render_json_schema_example(json_schema)
}
