% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openproblems.docs-package.R
\docType{package}
\name{openproblems.docs-package}
\alias{openproblems.docs}
\alias{openproblems.docs-package}
\title{openproblems.docs: OpenProblems Documentation}
\description{
OpenProblems Documentation R helper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openproblems.bio/documentation}
  \item \url{https://github.com/openproblems-bio/core}
  \item Report bugs at \url{https://github.com/openproblems-bio/core/issues}
}

}
\author{
\strong{Maintainer}: Kai Waldrant \email{k_waldant@hotmail.com} (\href{https://orcid.org/0009-0003-8555-1361}{ORCID})

}
\keyword{internal}
