% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_component_spec.R
\name{read_component_spec}
\alias{read_component_spec}
\title{read component spec}
\usage{
read_component_spec(path)
}
\arguments{
\item{path}{path to comp yaml}
}
\value{
a list with compontent info and arguments
}
\description{
read component spec
}
\examples{
path <- system.file(
  "extdata", "example_project", "api", "comp_method.yaml",
  package = "openproblems.docs"
)

read_component_spec(path)
}
