% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_format.R
\name{read_file_format}
\alias{read_file_format}
\title{Read file specification}
\usage{
read_file_format(path)
}
\arguments{
\item{path}{Path to yaml file}
}
\value{
list with file info and slots
}
\description{
Read file specification
}
\examples{
path <- system.file(
  "extdata", "example_project", "api", "file_train_h5ad.yaml",
  package = "openproblems.docs"
)

read_file_format(path)
}
