% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_format.R
\name{read_file_format}
\alias{read_file_format}
\title{Read file format}
\usage{
read_file_format(path)
}
\arguments{
\item{path}{Path to a file format yaml, usually in \code{src/api/file_*.yaml}}
}
\value{
A list with file format info and expected_format
}
\description{
This function reads a file format spec from a yaml file.
}
\examples{
path <- system.file(
  "extdata", "example_project", "api", "file_train_h5ad.yaml",
  package = "openproblems.docs"
)

read_file_format(path)
}
