% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_task_config.R
\name{read_task_config}
\alias{read_task_config}
\title{Read project config}
\usage{
read_task_config(file)
}
\arguments{
\item{file}{Path to a project config file}
}
\description{
Read project config
}
\examples{
file <- system.file(
  "extdata", "example_project", "_viash.yaml",
  package = "openproblems.docs"
)

read_task_config(file)
}
