% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_component_spec.R
\name{render_component_spec}
\alias{render_component_spec}
\title{Render component section}
\usage{
render_component_spec(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
string
}
\description{
Render component section
}
\examples{
path <- system.file(
  "extdata", "example_project", "api", "comp_method.yaml",
  package = "openproblems.docs"
)

spec <- read_component_spec(path)

render_component_spec(spec)
}
