% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_file.R
\name{render_file_format}
\alias{render_file_format}
\title{Render file section}
\usage{
render_file_format(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
string
}
\description{
Render file section
}
\examples{
path <- system.file("extdata", "example_project", "api", "file_train_h5ad.yaml", package = "openproblems.docs")

render_component_spec(path)
}
