% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_task_readme_qmd.R
\name{render_task_readme_qmd}
\alias{render_task_readme_qmd}
\title{Render the README.qmd file for a task}
\usage{
render_task_readme_qmd(task_metadata, add_instructions = FALSE)
}
\arguments{
\item{task_metadata}{Task metadata as returned by \code{read_task_metadata}}

\item{add_instructions}{Add instructions to the README.qmd}
}
\value{
A qmd documentation string
}
\description{
Render the README.qmd file for a task
}
\examples{
path <- system.file("extdata", "example_project", package = "openproblems.docs")

task_metadata <- read_task_metadata(path)

render_task_readme_qmd(task_metadata)
}
